HACK: Hardcode hdmi_port as HDMI3, base_device as Tv
Oddly enough, both of these apparently need to be set for the Pulse-Eight to get
the correct HDMI3 with physical address 3.0.0.0 attached to it.  Without these,
the device would default to HDMI1 / 1.0.0.0 and set active source to the wrong
input!

If base_device is `CecLogicalAddress::Unknown`, then the HDMI address defaults
to HDMI1 / 1.0.0.0 regardless of `hdmi_port` being set. It seems this is a
trigger for the autodetection logic, which seems to fail in all cases where a PC
has multiple HDMI ports and the Pulse-Eight adapter is not connected to the very
first one.

Likewise, if `hdmi_port` is unset, but `base_device` is either
`CecLogicalAddress:Tv` or unset, it also defaults to HDMI1 / 1.0.0.0 improperly.

Ironically, setting `physical_address` to
`libcec_sys::CEC_INVALID_PHYSICAL_ADDRESS` does not break anything so long as
`hdmi_port` is set and `base_device` is not `Unknown` also.

Adding this hardcoded hack for now just so things will work and be stable again.