HACK: Hardcode hdmi_port as HDMI3, base_device as Tv
Oddly enough, both of these apparently need to be set for the Pulse-Eight to get
the correct HDMI3 with physical address 3.0.0.0 attached to it.  Without these,
the device would default to HDMI1 / 1.0.0.0 and set active source to the wrong
input!

If base_device is `CecLogicalAddress::Unknown`, then the HDMI address defaults
to HDMI1 / 1.0.0.0 regardless of `hdmi_port` being set. It seems this is a
trigger for the autodetection logic, which seems to fail in all cases where a PC
has multiple HDMI ports and the Pulse-Eight adapter is not connected to the very
first one.

Likewise, if `hdmi_port` is unset, but `base_device` is either
`CecLogicalAddress:Tv` or unset, it also defaults to HDMI1 / 1.0.0.0 improperly.

Ironically, setting `physical_address` to
`libcec_sys::CEC_INVALID_PHYSICAL_ADDRESS` does not break anything so long as
`hdmi_port` is set and `base_device` is not `Unknown` also.

Adding this hardcoded hack for now just so things will work and be stable again.

It turns out that `libcec` looks for the first `enabled` video card device,
then reads the EDID data from that to try and detect a physical CEC address.

However, this is not always the right choice for systems with multiple video cards,
and multiple active HDMI outputs going to different monitors or TVs.

TODO: Rework this into a config file to give `libcec` the proper HDMI port and/or
physical address  hints.

Reference:

  - https://github.com/Pulse-Eight/libcec/blob/95d1c6965e38000f56776f4563f9763832a5a7c2/src/libcec/platform/drm/drm-edid.cpp#L47-L81