config: Fix failing capacity overflow test for device_types deserializer
It turns out `break` wasn't the proper way to avoid panic on deserialization of
a YAML array larger than the `ArrayVec` capacity.

Refactored this deserializer to fully consume the input and satisfy serde's
expectations.  Thus, we never panic on oversized inputs.

This is a more lenient config file deserializer, to warn the user but still avoid
a panic.