config: Refactor CecDpmsConfig into two structs: CecDpms{Root,Adapter}Config
This allows the config.yaml file schema to support multiple CEC adapters, each
with its own physical address and connection settings.

Note: This _would be_ a breaking change to the YAML file schema, except that we
have not yet released the code supporting a config file at all.

As such, this is a WIP feature under flux until release.